refactor: move chunk infrastructure to service/chunk/ package
- Move ChunkService + RetrievalTest + hydrateChunkVectors to service/chunk/
- Move FetchChunkVectors to service/chunk/vector.go
- Move PtrString to common/format.go (generic utility)
- Merge chunk_view.go into chunk_types.go (type definitions)
- Rename chunk_view_test.go → chunk_types_test.go
- Split KbPrompt tests → kb_prompt_test.go
- Move hydrate/isZero tests to service/chunk/chunk_test.go
- Remove duplicate chunk_vector files from service/
- service/chunk/ imports service/ only (no circular dependency)
