+++ cmd/server_main.go
@@ -40,6 +40,7 @@ import (
 	"ragflow/internal/dao"
 	"ragflow/internal/engine"
 	"ragflow/internal/handler"
+	"ragflow/internal/service/chunk"
 	"ragflow/internal/router"
 	"ragflow/internal/service"
 	"ragflow/internal/service/nlp"
@@ -71,7 +72,7 @@ func main() {
 
 	// Initialize logger with default level
 	// logger.Init("info"); // set debug log level
-	if err := common.Init("info"); err != nil {
+	if err := common.Init("info", "server_main.log"); err != nil {
 		panic(fmt.Sprintf("Failed to initialize logger: %v", err))
 	}
 
@@ -92,10 +93,12 @@ func main() {
 	}
 
 	// Reinitialize logger with configured level if different
-	if config.Log.Level != "" && config.Log.Level != "info" {
-		if err := common.Init(config.Log.Level); err != nil {
-			common.Error("Failed to reinitialize logger with configured level", err)
-		}
+	level := config.Log.Level
+	if level == "" {
+		level = "info"
+	}
+	if err := common.Init(level, "server_main.log"); err != nil {
+		common.Error("Failed to reinitialize logger", err)
 	}
 	server.SetLogger(common.Logger)
 	if config.Log.Level == "" {
@@ -176,7 +179,7 @@ func startServer(config *server.Config) {
 	datasetsService := service.NewDatasetService()
 	knowledgebaseService := service.NewKnowledgebaseService()
 	metadataService := service.NewMetadataService()
-	chunkService := service.NewChunkService()
+	chunkService := chunk.NewChunkService()
 	llmService := service.NewLLMService()
 	tenantService := service.NewTenantService()
 	chatService := service.NewChatService()
@@ -224,6 +227,7 @@ func startServer(config *server.Config) {
 	searchBotHandler.SetAskService(service.NewAskService(chunkService, nil, 0, 0))
 	searchBotHandler.SetStreamLLM(searchBotLLM)
 	searchBotHandler.SetAskService(service.NewAskService(chunkService, nil, 0, 0))
+	pluginHandler := handler.NewPluginHandler(service.NewPluginService())
 
 	// Dify retrieval handler
 	docDAO := dao.NewDocumentDAO()
