+++ internal/handler/chunk.go
@@ -1,4 +1,3 @@
-//
 //  Copyright 2026 The InfiniFlow Authors. All Rights Reserved.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
