+++ internal/service/chunk/vector_test.go
@@ -14,14 +14,14 @@
 //  limitations under the License.
 //
 
-package service
+package chunk
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"reflect"
 	"testing"
+	"context"
 
 	"ragflow/internal/engine/types"
 )
