+++ internal/service/chunk_types_test.go
@@ -17,9 +17,6 @@
 package service
 
 import (
-	"context"
-
-	"reflect"
 	"testing"
 )
 
@@ -160,99 +157,15 @@ func TestChunksFormat_FieldMapping(t *testing.T) {
 	}
 }
 
-func TestKbPrompt_Empty(t *testing.T) {
-	if got := KbPrompt(nil, 100); got != "" {
-		t.Errorf("expected empty for nil chunks")
+func TestGetMap_NilAndMissing(t *testing.T) {
+	if getMap(nil, "x") != nil {
+		t.Error("nil map should return nil")
 	}
-	if got := KbPrompt([]SourcedChunk{}, 100); got != "" {
-		t.Errorf("expected empty for empty chunks")
+	if getMap(map[string]interface{}{}, "x") != nil {
+		t.Error("missing key should return nil")
 	}
-	if got := KbPrompt([]SourcedChunk{{Content: "x"}}, 0); got != "" {
-		t.Errorf("expected empty for maxTokens=0")
-	}
-	if got := KbPrompt([]SourcedChunk{{Content: "x"}}, -1); got != "" {
-		t.Errorf("expected empty for maxTokens=-1")
-	}
-}
-
-func TestKbPrompt_Format(t *testing.T) {
-	chunks := []SourcedChunk{{
-		ID:      "abc",
-		Content: "chunk content here",
-		DocName: "Test Document",
-		URL:     "http://example.com",
-	}}
-	result := KbPrompt(chunks, 10000)
-	if result == "" {
-		t.Fatal("expected non-empty prompt")
-	}
-	// Verify ID appears
-	if !contains(result, "ID: abc") {
-		t.Errorf("missing ID line: %s", result)
-	}
-	// Verify title
-	if !contains(result, "Title: Test Document") {
-		t.Errorf("missing title: %s", result)
-	}
-	// Verify URL
-	if !contains(result, "URL: http://example.com") {
-		t.Errorf("missing URL: %s", result)
-	}
-	// Verify content
-	if !contains(result, "chunk content here") {
-		t.Errorf("missing content: %s", result)
-	}
-	// Verify unicode box-drawing chars
-	if !contains(result, "├──") {
-		t.Errorf("missing tree drawing: %s", result)
-	}
-}
-
-func TestKbPrompt_TokenLimit(t *testing.T) {
-	chunks := []SourcedChunk{
-		{ID: "1", Content: "a very long content that takes many tokens "},
-		{ID: "2", Content: "second chunk content here"},
-	}
-	// Tight limit: first chunk ~31 tokens (limit=48 tokens at 0.97 ratio).
-	// Second chunk ~25 tokens — excluded.
-	result := KbPrompt(chunks, 50)
-	if !contains(result, "ID: 1") {
-		t.Error("first chunk should be included")
-	}
-	if contains(result, "ID: 2") {
-		t.Error("second chunk should be excluded under tight limit")
-	}
-}
-
-func TestKbPrompt_DocMetadata(t *testing.T) {
-	chunks := []SourcedChunk{{
-		ID:      "abc",
-		Content: "content",
-		DocumentMetadata: map[string]interface{}{
-			"author": "test author",
-			"year":   "2024",
-		},
-	}}
-	result := KbPrompt(chunks, 10000)
-	if !contains(result, "author: test author") {
-		t.Errorf("missing metadata author: %s", result)
-	}
-	if !contains(result, "year: 2024") {
-		t.Errorf("missing metadata year: %s", result)
-	}
-}
-
-func TestKbPrompt_NoDocNameOrURL(t *testing.T) {
-	chunks := []SourcedChunk{{
-		ID:      "simple",
-		Content: "plain content",
-	}}
-	result := KbPrompt(chunks, 10000)
-	if contains(result, "Title:") {
-		t.Error("should not have title when empty")
-	}
-	if contains(result, "URL:") {
-		t.Error("should not have URL when empty")
+	if getMap(map[string]interface{}{"x": "not a map"}, "x") != nil {
+		t.Error("wrong type should return nil")
 	}
 }
 
@@ -292,15 +205,6 @@ func TestGetFloat_Types(t *testing.T) {
 	if getFloat(m, "missing") != 0 {
 		t.Error("missing should return 0")
 	}
-}
-
-func contains(s, substr string) bool {
-	for i := 0; i <= len(s)-len(substr); i++ {
-		if s[i:i+len(substr)] == substr {
-			return true
-		}
-	}
-	return false
 }
 
 func TestSourcedChunk_ZeroValue(t *testing.T) {
@@ -349,120 +253,49 @@ func TestNewSourcedChunks_RoundTrip(t *testing.T) {
 	}
 }
 
-func TestNumTokensFromString_Empty(t *testing.T) {
-	if got := NumTokensFromString(""); got != 0 {
-		t.Errorf("expected 0 for empty string, got %d", got)
+func TestGetPageNum_Nil(t *testing.T) {
+	if got := getPageNum(nil, 10); got != 10 {
+		t.Errorf("getPageNum(nil, 10) = %d, want 10", got)
 	}
 }
 
-func TestNumTokensFromString_Fallback(t *testing.T) {
-	// With tokenizer unavailable, fallback to rune count / 2.
-	s := "hello world"
-	got := NumTokensFromString(s)
-	expected := len([]rune(s)) / 2 // 5
-	if got != expected {
-		t.Errorf("got %d, want %d (fallback)", got, expected)
+func TestGetPageNum_ZeroReturnsDefault(t *testing.T) {
+	val := 0
+	if got := getPageNum(&val, 5); got != 5 {
+		t.Errorf("getPageNum(&0, 5) = %d, want 5", got)
 	}
 }
 
-func TestNumTokensFromString_Chinese(t *testing.T) {
-	s := "你好世界"
-	got := NumTokensFromString(s)
-	expected := len([]rune(s)) / 2 // 2
-	if got != expected {
-		t.Errorf("got %d, want %d (Chinese fallback)", got, expected)
+func TestGetPageNum_NegativeReturnsDefault(t *testing.T) {
+	val := -1
+	if got := getPageNum(&val, 5); got != 5 {
+		t.Errorf("getPageNum(&-1, 5) = %d, want 5", got)
 	}
 }
 
-func TestKbPrompt_TokenLimitAccurate(t *testing.T) {
-	// Verify truncation uses NumTokensFromString, not byte length.
-	chunks := []SourcedChunk{
-		{ID: "1", Content: "hello"},   // ~10 runes → 5 tokens + overhead
-		{ID: "2", Content: "world"},   // ~5 tokens
-	}
-	// With maxTokens=20, limit=19→ first fits, second doesn't.
-	result := KbPrompt(chunks, 20)
-	if !contains(result, "ID: 1") {
-		t.Error("first chunk should fit under 20 token limit")
-	}
-	if contains(result, "ID: 2") {
-		t.Errorf("second chunk should be excluded: result = %q", result)
+func TestGetPageNum_Valid(t *testing.T) {
+	val := 3
+	if got := getPageNum(&val, 5); got != 3 {
+		t.Errorf("getPageNum(&3, 5) = %d, want 3", got)
 	}
 }
 
-func TestKbPrompt_AllFit(t *testing.T) {
-	chunks := []SourcedChunk{
-		{ID: "1", Content: "a"},
-		{ID: "2", Content: "b"},
-	}
-	result := KbPrompt(chunks, 1000)
-	if !contains(result, "ID: 1") && !contains(result, "ID: 2") {
-		t.Error("both chunks should fit under generous limit")
+func TestGetPageSize_Nil(t *testing.T) {
+	if got := getPageSize(nil, 20); got != 20 {
+		t.Errorf("getPageSize(nil, 20) = %d, want 20", got)
 	}
 }
 
-func TestIsZeroVector(t *testing.T) {
-	if !isZeroVector([]float64{0, 0, 0}) {
-		t.Error("all zeros should be true")
-	}
-	if isZeroVector([]float64{0, 1, 0}) {
-		t.Error("non-zero should be false")
-	}
-	if !isZeroVector([]float64{}) {
-		t.Error("empty should be true (treated as zero)")
-	}
-	if !isZeroVector(nil) {
-		t.Error("nil should be true")
+func TestGetPageSize_ZeroReturnsDefault(t *testing.T) {
+	val := 0
+	if got := getPageSize(&val, 20); got != 20 {
+		t.Errorf("getPageSize(&0, 20) = %d, want 20", got)
 	}
 }
 
-func TestHydrateChunkVectors_AllNonZero(t *testing.T) {
-	chunks := []map[string]interface{}{
-		{"id": "c1", "vector": []float64{1, 2, 3}},
-		{"id": "c2", "vector": []float64{4, 5, 6}},
-	}
-	// No zero vectors → nothing to hydrate.
-	hydrateChunkVectors(context.Background(), nil, chunks, nil, nil)
-	if !reflect.DeepEqual(chunks[0]["vector"], []float64{1, 2, 3}) {
-		t.Error("non-zero vector should not be changed")
-	}
-	if !reflect.DeepEqual(chunks[1]["vector"], []float64{4, 5, 6}) {
-		t.Error("non-zero vector should not be changed")
-	}
-}
-
-func TestHydrateChunkVectors_EmptyChunks(t *testing.T) {
-	// Should not panic on empty or nil.
-	hydrateChunkVectors(context.Background(), nil, nil, nil, nil)
-	hydrateChunkVectors(context.Background(), nil, []map[string]interface{}{}, nil, nil)
-}
-
-func TestHydrateChunkVectors_MissingIDs(t *testing.T) {
-	chunks := []map[string]interface{}{
-		{"vector": []float64{1.0}}, // no id — skipped
-	}
-	hydrateChunkVectors(context.Background(), nil, chunks, nil, nil)
-	// Should not change anything when engine is nil (FetchChunkVectors returns zero vectors).
-	// The function doesn't panic — it just can't hydrate because dim is 0.
-	// With nil engine, FetchChunkVectors returns zero vectors, so the zero stays zero.
-}
-
-func TestHydrateChunkVectors_NoDim(t *testing.T) {
-	chunks := []map[string]interface{}{
-		{"id": "c1", "vector": []float64{}},
-	}
-	hydrateChunkVectors(context.Background(), nil, chunks, []string{"kb1"}, []string{"t1"})
-	// Empty vectors have dim=0 → early return. No crash.
-}
-
-func TestGetMap_NilAndMissing(t *testing.T) {
-	if getMap(nil, "x") != nil {
-		t.Error("nil map should return nil")
-	}
-	if getMap(map[string]interface{}{}, "x") != nil {
-		t.Error("missing key should return nil")
-	}
-	if getMap(map[string]interface{}{"x": "not a map"}, "x") != nil {
-		t.Error("wrong type should return nil")
+func TestGetPageSize_Valid(t *testing.T) {
+	val := 50
+	if got := getPageSize(&val, 20); got != 50 {
+		t.Errorf("getPageSize(&50, 20) = %d, want 50", got)
 	}
 }
